requirements.txt
PyYAML>=6.0
jsonschema>=4.0
jsonschema>=3.2.0
